Create LinkedList#at
This method uses recursion to find and return the node at the specified index given. It assumes the linked list to be 0 index.

The #append method was modified to set the head and the tail of the linked list to the same node if it is the sole node of the linked list.