Create LinkedList#insert_at
This method takes two arguments, value and index, and inserts the value at the index while rerouting the linked list as necessary.

It will throw an error if a non-existing index is provided, i.e., the highest index that can be provided for this method to run correctly is #size + 1.