Create LinkedList#delete_at
This method has a required parameter, index, and removes the node at the index of the linked list from the linked list. It also accounts for the reduced size of the linked list.

The #insert_at method was modified to account for the linked list size change and return the inserted node.